+++ src/DevilDaggersInfo.Core.Mod/ShaderData.cs
@@ -1,0 +1,3 @@
+namespace DevilDaggersInfo.Core.Mod;
+
+internal record ShaderData(string ShaderName, byte[] VertexShaderFileContents, byte[] FragmentShaderFileContents);