+++ src/test/DevilDaggersInfo.Core.Mod.Test/CompilationTests.cs
@@ -1,4 +1,6 @@
 using DevilDaggersInfo.Core.Asset;
+using DevilDaggersInfo.Core.Mod.Builders;
+using DevilDaggersInfo.Core.Mod.Exceptions;
 
 namespace DevilDaggersInfo.Core.Mod.Test;
 
@@ -11,12 +13,48 @@ public class CompilationTests
 	{
 		byte[] sourcePngContents = File.ReadAllBytes(Path.Combine("Resources", "Texture", sourcePngFileName));
 
-		ModBinaryBuilder modBinary = new(ModBinaryType.Dd);
-		modBinary.AddAsset(assetName, AssetType.Texture, sourcePngContents);
+		DdModBinaryBuilder builder = new();
+		builder.AddTexture(assetName, sourcePngContents);
 
-		byte[] compiledModBinary = modBinary.Compile();
+		byte[] compiledModBinary = builder.Compile();
 
 		byte[] sourceModBinary = File.ReadAllBytes(Path.Combine("Resources", modFileName));
 		CollectionAssert.AreEqual(compiledModBinary, sourceModBinary);
+	}
+
+	[DataTestMethod]
+	[DataRow("depth.vert", "depth.frag", "depth")]
+	public void CompileShaderIntoModBinary(string sourceVertexFileName, string sourceFragmentFileName, string assetName)
+	{
+		byte[] sourceVertexContents = File.ReadAllBytes(Path.Combine("Resources", "Shader", sourceVertexFileName));
+		byte[] sourceFragmentContents = File.ReadAllBytes(Path.Combine("Resources", "Shader", sourceFragmentFileName));
+
+		DdModBinaryBuilder modBinary = new();
+		modBinary.AddShader(assetName, sourceVertexContents, sourceFragmentContents);
+
+		byte[] compiledModBinary = modBinary.Compile();
+
+		ModBinary extractedModBinary = new(compiledModBinary, ModBinaryReadFilter.AllAssets);
+		AssetExtractionResult assetExtractionResult = extractedModBinary.ExtractAsset(assetName, AssetType.Shader);
+		Assert.AreEqual(2, assetExtractionResult.ExtractedAssetFiles.Count);
+
+		if (assetExtractionResult.ExtractedAssetFiles.TryGetValue($"{assetName}.vert", out byte[]? vertexContents))
+			CollectionAssert.AreEqual(vertexContents, sourceVertexContents);
+		else
+			Assert.Fail("Vertex shader not found in extracted asset files.");
+
+		if (assetExtractionResult.ExtractedAssetFiles.TryGetValue($"{assetName}.frag", out byte[]? fragmentContents))
+			CollectionAssert.AreEqual(fragmentContents, sourceFragmentContents);
+		else
+			Assert.Fail("Fragment shader not found in extracted asset files.");
+	}
+
+	[TestMethod]
+	public void TestDuplicateAsset()
+	{
+		DdModBinaryBuilder builder = new();
+		builder.AddObjectBinding("test", Array.Empty<byte>());
+		builder.AddMesh("test", Array.Empty<byte>());
+		Assert.ThrowsException<InvalidModCompilationException>(() => builder.AddObjectBinding("test", Array.Empty<byte>()));
 	}
 }